// src/Api/group/dto/group-charge.dto.ts
export class GroupChargeDto {
  terminal: string;
  group: string;
  amount: number;
  expire: string;
  users: string;
  readonly saturday: boolean;
  readonly sunday: boolean;
  readonly monday: boolean;
  readonly tuesday: boolean;
  readonly wednesday: boolean;
  readonly thursday: boolean;
  readonly friday: boolean;
  title: string;
}