// src/Api/group/dto/group-organization.dto.ts
export class OrganizationGroupStrategyDto {
  group: string;
  terminal: string;
  min: number;
  max: number;
  expire: string;
  refresh: number;
  amount: number;
  users: string;
  customercharge: number;
  saturday: boolean;
  sunday: boolean;
  monday: boolean;
  tuesday: boolean;
  wednesday: boolean;
  thursday: boolean;
  friday: boolean;
}